Update aliases, insert date in journal
This removes old aliases that were redundant or that I never got
in the habit of using. I also add the date and time to my journal
in my preferred format before opening it. This appears to be
fairly robust, regardless of whether I add a newline at the end
of my last line before saving the journal. My guess is that's
because the last line only contains an EOF character, so I would
need a newline regardless to get my expected spacing. "Yes, self,
it would be a good idea to post your guess to GitHub."