Use journal-date, printf with jr instead of emacs commands
This uses the Python script and some printf statements to set up
journal file for editing. The nice thing is this makes it more
generic: I can switch from emacs if I want (not sure I do, but I
am playing with vim lately). The minor annoyance is that if I type
"jr" but don't end up writing anything, the file is changed (a
timestamp is added). I sometimes do this and abandon changes, but
it's more out of boredom than anything. I'm not sure if the
flexibility is worth it, but I do want to hang on to these changes
so I will commit them. (It's already annoying to run jr, abandon
changes, and have emacs try to recover the auto-save, anyway.) I
like flexibility even though the freest of free software with tons
of support is not going anywhere anytime soon. Then again, doing
it all in emacs probably works on Windows, or could be made to
work more easily than this chain of bash commands. Except you can
run bash on Windows now...