Refactor jr for cleaner Mac ~/.bash_profile
This change creates a new emacs function, journal-date, that does
the checking the if statement in my bash_profile used to do. It
seems like a more reasonable place to contain the logic. Of course,
the last commit was intended to make me less reliant on emacs so
maybe that's enough computer for today.