Refactor jr to use vim, Python date script
This code is more explicit / clear about what it does. While emacs
handles things like folder creation (at least at saving time), this
creates one if it's needed, adds extra newlines if the file exists,
and opens up in vim. I'm on a vim kick lately so this is a good
thing, although my git editor is still emacs, which just feels
right for some reason. Perhaps it's the complicated keystrokes to
save.