Switch to iA Writer for journal (jr)
This is more about switching away from Dropbox to iCloud than it is
about preferring iA Writer to vim for my journal. I was already
using iA Writer on my phone and tablet. I think I CAN use vim with
the Mobile Documents folder or whatever it's called where iCloud
actually syncs on disk. But that makes me a little nervous since
the file paths are so weird. Like, would it work? Would it stay in
sync? I assume it would, but I'm not sure.

Dropbox is annoying. To be fair, I used their free version for years
on a lot of computers. But the app seems to suck power on my Mac and
I suspect it's the reason my battery percentage drops when my
laptop is supposed to be asleep. Guess I'll find out soon.