Add minor updates to VSCode settings
I think editor.folding is so the text in a new window isn't indented
so weirdly far in. Otherwise, I’m preferring to open files in new
tabs, rather than windows, as I’ve noticed much of what makes VS
Code feel slow to me is how long it takes to open a window on a Mac.